+++ R/nonlinear_ISPE.R
@@ -86,8 +86,8 @@ do.ispe <- function(X, ndim=2, proximity=function(x){dist(x, method="euclidean")
   ## MAIN COMPUTATION
   # 1. Proximity Matrix R : don't forget to make it into a matrix.
   R = proximity(X)
-  if ((inherits(R, "dist"))||(class(R)=="dist")){
-    R = as.matrix(R)
+  if (inherits(R,"dist")){
+    R = as.matrix
   }
   # 2. initial coordinates
   iX = do.pca(X, ndim=ndim)$Y